+++ Escuela/libft/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: rarias-p <rarias-p@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/06 09:30:04 by rarias-p          #+#    #+#             */
-/*   Updated: 2019/11/15 09:23:30 by rarias-p         ###   ########.fr       */
+/*   Updated: 2019/12/19 10:35:05 by rarias-p         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *str)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	while (str[i] != '\0')
